restructure directories for further development
