--- src/components/Button/Button.module.scss
@@ -1,19 +1,0 @@
-@import "../../styles/vars.scss";
-
-.button {
-  background-color: $akai-yellow;
-  border: 0;
-  border-radius: 25px;
-  color: $akai-white;
-  cursor: pointer;
-  font-size: 15px;
-  font-family: inherit;
-  font-weight: 600;
-  margin: 0 auto;
-  padding: 15px 45px;
-  transition: background-color 0.3s ease-out;
-
-  &:hover {
-    background-color: $akai-yellow-darkened;
-  }
-}
