--- src/components/Icon/index.js
@@ -1,21 +1,0 @@
-import Icon from "./Icon.jsx";
-import {
-  mdiNewspaperVariantOutline,
-  mdiApps,
-  mdiCalendar,
-  mdiFacebook,
-  mdiTwitter,
-  mdiGithub,
-  mdiLinkedin,
-} from "@mdi/js";
-
-export {
-  Icon as default,
-  mdiNewspaperVariantOutline as blog,
-  mdiApps as apps,
-  mdiCalendar as calendar,
-  mdiFacebook as facebook,
-  mdiTwitter as twitter,
-  mdiGithub as github,
-  mdiLinkedin as linkedin,
-};
