+++ src/components/MainPage/index.js
@@ -1,0 +1,6 @@
+export { default as Main } from "./Main/Main";
+export { default as About } from "./About/About";
+export { default as LinkIcons } from "./LinkIcons/LinkIcons";
+export { default as JoinUs } from "./JoinUs/JoinUs";
+export { default as Partners } from "./Partners/Partners";
+export { default as Footer } from "./Footer/Footer";
