--- src/components/Section/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Section.jsx";
