--- src/components/Title/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Title.jsx";
