+++ src/components/ui/Icon/Icon.module.scss
