+++ src/components/ui/Icon/index.js
@@ -1,0 +1,19 @@
+import {
+  mdiNewspaperVariantOutline,
+  mdiApps,
+  mdiCalendar,
+  mdiFacebook,
+  mdiTwitter,
+  mdiGithub,
+  mdiLinkedin,
+} from "@mdi/js";
+
+export {
+  mdiNewspaperVariantOutline as blog,
+  mdiApps as apps,
+  mdiCalendar as calendar,
+  mdiFacebook as facebook,
+  mdiTwitter as twitter,
+  mdiGithub as github,
+  mdiLinkedin as linkedin,
+};
