+++ src/components/ui/Title/Title.module.scss
@@ -1,0 +1,36 @@
+@import "../../../styles/vars.scss";
+
+.title {
+  font-size: 25px;
+  margin: 15px 0;
+
+  &::after {
+    content: "";
+    display: block;
+    background-color: $akai-yellow;
+    height: 3px;
+    width: 40px;
+  }
+}
+
+.center {
+  text-align: center;
+
+  &::after {
+    margin: 15px auto 25px;
+  }
+}
+
+.left {
+  @extend .center;
+}
+
+@media only screen and (min-width: $tablet) {
+  .left {
+    text-align: left;
+
+    &::after {
+      margin: 15px 0 25px;
+    }
+  }
+}
