+++ src/components/ui/index.js
@@ -1,0 +1,5 @@
+export { default as Button } from "./Button/Button";
+export { default as Icon } from "./Icon/Icon";
+export { default as Navigation } from "./Navigation/Navigation";
+export { default as Section } from "./Section/Section";
+export { default as Title } from "./Title/Title";
