+++ src/pages/index.js
@@ -1,11 +1,13 @@
 import React from "react";
-import SEO from "../components/SEO";
-import Main from "../sections/Main";
-import About from "../sections/About";
-import LinkIcons from "../sections/LinkIcons";
-import JoinUs from "../sections/JoinUs";
-import Footer from "../sections/Footer";
-import Partners from "../sections/Partners/Partners";
+import SEO from "@/components/SEO";
+import {
+  About,
+  JoinUs,
+  LinkIcons,
+  Main,
+  Partners,
+  Footer,
+} from "@/components/MainPage";
 
 export default function Home() {
   return (
