--- src/sections/About/index.js
@@ -1,1 +1,0 @@
-export { default } from "./About.jsx";
