--- src/sections/Footer/Footer.module.scss
@@ -1,72 +1,0 @@
-@import "./../../styles/vars.scss";
-
-.section {
-  color: $heading-color;
-  display: flex;
-  text-align: center;
-  flex-direction: column-reverse;
-}
-
-.heading {
-  font-size: 15px;
-  margin: 15px 0;
-}
-
-.copyright {
-  flex-basis: 50%;
-}
-
-.contact {
-  flex-basis: 30%;
-
-  a {
-    color: $heading-color;
-  }
-}
-
-.social {
-  flex-basis: 20%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.socialIcons {
-  display: flex;
-
-  a + a {
-    margin-left: 10px;
-  }
-}
-
-.iconWrapper {
-  padding: 7px;
-  border: 1px solid $text-color;
-  border-radius: 50%;
-  fill: $text-color;
-  transition: fill $hoverTransitionDuration ease-in-out,
-    border-color $hoverTransitionDuration ease-in-out;
-
-  &:hover {
-    fill: $text-color-hover;
-    border-color: $text-color-hover;
-  }
-}
-
-.icon {
-  width: 24px;
-  height: 24px;
-  display: flex;
-  fill: inherit;
-}
-
-@media only screen and (min-width: $tablet) {
-  .section {
-    flex-direction: row;
-    text-align: left;
-  }
-
-  .social {
-    align-items: flex-start;
-  }
-}
