--- src/sections/JoinUs/index.js
@@ -1,1 +1,0 @@
-export { default } from "./JoinUs.jsx";
