--- src/sections/LinkIcons/LinkIcons.module.scss
@@ -1,73 +1,0 @@
-@import "../../styles/vars.scss";
-
-.section {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  margin: auto;
-  max-width: 640px;
-
-  a {
-    text-decoration: none;
-  }
-}
-
-.iconsWrapper {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  flex-wrap: wrap;
-}
-
-.iconLink {
-  flex-basis: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 20px;
-
-  &:hover {
-    .icon {
-      fill: $akai-yellow-darkened;
-    }
-
-    .text {
-      color: $heading-color-hover;
-    }
-
-    .description {
-      color: $text-color-hover;
-    }
-  }
-}
-
-.icon {
-  width: 90px;
-  height: 90px;
-  margin: 20px 20px 15px;
-  fill: $akai-yellow;
-  transition: fill $hoverTransitionDuration ease-in-out;
-}
-
-.text {
-  font-size: 25px;
-  font-weight: 600;
-  color: $heading-color;
-  margin-bottom: 10px;
-  transition: color $hoverTransitionDuration ease-in-out;
-}
-
-.description {
-  font-size: 15px;
-  font-weight: 500;
-  color: $text-color;
-  text-align: center;
-  max-width: 200px;
-  transition: color $hoverTransitionDuration ease-in-out;
-}
-
-@media only screen and (min-width: $mobile) {
-  .iconLink {
-    flex-basis: 33%;
-  }
-}
