--- src/sections/LinkIcons/index.js
@@ -1,1 +1,0 @@
-export { default } from "./LinkIcons.jsx";
