--- src/sections/Main/Main.module.scss
@@ -1,44 +1,0 @@
-@import "../../styles/vars.scss";
-
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  height: 100vh;
-  background: #1e2633 url("/img/background.jpg");
-  background-size: cover;
-  color: $akai-white;
-  text-align: center;
-}
-
-.top {
-  width: 100%;
-  padding: 30px 20px 0;
-  max-width: 1100px;
-  display: flex;
-  align-items: flex-start;
-  img {
-    width: 40px;
-  }
-}
-
-.title {
-  font-size: 48px;
-
-  &::after {
-    content: "";
-    display: block;
-    background-color: $akai-yellow;
-    height: 3px;
-    width: 40px;
-    margin: 20px auto;
-  }
-}
-
-.text {
-  max-width: 700px;
-  line-height: 2;
-  margin: 0 20px;
-  color: $akai-white;
-}
