--- src/sections/Partners/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Partners.jsx";
